requirements.txt
pytz
requests
python-dotenv
twine
setuptools